requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
boto3==1.28.56
botocore==1.31.56
celery==5.3.4
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.2.0
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
distlib==0.3.7
dj-config-url==0.1.1
dj-database-url==2.1.0
Django==4.2.5
django-redis==5.3.0
djangorestframework==3.14.0
drf-yasg==1.21.7
filelock==3.12.4
frozenlist==1.4.0
future==0.18.3
fuzzywuzzy==0.18.0
identify==2.5.30
idna==3.4
inflection==0.5.1
jmespath==1.0.1
kombu==5.3.2
multidict==6.0.4
nodeenv==1.8.0
packaging==23.1
phonenumbers==8.13.21
platformdirs==3.10.0
pre-commit==3.4.0
prompt-toolkit==3.0.39
psycopg2==2.9.8
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python-http-client==3.3.7
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.0
requests==2.31.0
s3transfer==0.7.0
sendgrid==6.10.0
serializers==0.2.4
six==1.16.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
twilio==8.9.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
virtualenv==20.24.5
wcwidth==0.2.8
yarl==1.9.2
